fix(x86/klib): fix timing and printf issues by correcting instruction logic and klib impl - nemu: fix sbb/adc macros where write-back occurred before flag calculation - klib: enhance printf to support width and zero-padding (e.g., %02d) - klib: fix __udivdi3 to ensure EDX is cleared for 32-bit result paths - nemu: fix potential 32-bit overflow in host timer calculation These changes fix the issue where am-tests rtc_test would output too fast and display mangled format specifiers.

+++ nemu/src/utils/timer.c
@@ -29,11 +29,11 @@ static uint64_t get_time_internal() {
 #elif defined(CONFIG_TIMER_GETTIMEOFDAY)
   struct timeval now;
   gettimeofday(&now, NULL);
-  uint64_t us = now.tv_sec * 1000000 + now.tv_usec;
+  uint64_t us = (uint64_t)now.tv_sec * 1000000 + now.tv_usec;
 #else
   struct timespec now;
   clock_gettime(CLOCK_MONOTONIC_COARSE, &now);
-  uint64_t us = now.tv_sec * 1000000 + now.tv_nsec / 1000;
+  uint64_t us = (uint64_t)now.tv_sec * 1000000 + now.tv_nsec / 1000;
 #endif
   return us;
 }
